+++ CHANGELOG.md
@@ -6,3 +6,10 @@
 - StarConfig class for configuring star appearance and behavior
 - StarAnimationConfig class for configuring star animations
 - StarEnums for star layout direction, size variants, star animation types, filling behavior, alignment, interaction mode, and more
+
+## 1.0.1
+
+* Bug fixes and improvements
+* Added emoji rating widget
+* Added support for animated emoji slider
+* Added support for custom emojis